# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Documentación generada con Middleware a Hubspot",
    "description": "Documentación proyecto de Contactos Trainee"
  },
  "host": "trainee.up.railway.app",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/contacts/": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "description": "Get all contacts service",
        "responses": {
          "200": {
            "description": "Get all users",
            "schema": {
              "$ref": "#/definitions/Contacts"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Contacts"
        ],
        "description": "POST a new contact service",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "firstname": {
                  "example": "any"
                },
                "lastname": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post a new Contact",
            "schema": {
              "$ref": "#/definitions/Contacts"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/contacts/{mail}": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "description": "Get contact that match email",
        "parameters": [
          {
            "name": "mail",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Get user that contains email",
            "schema": {
              "$ref": "#/definitions/Contacts"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/contacts/{cid}": {
      "patch": {
        "tags": [
          "Contacts"
        ],
        "description": "PATCH a contact by ID service",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "firstname": {
                  "example": "any"
                },
                "lastname": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Update contact data that match id ",
            "schema": {
              "$ref": "#/definitions/Contacts"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Contacts"
        ],
        "description": "Delete a contact by ID service",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Delete a Contact"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "Contacts": {
      "type": "object",
      "properties": {
        "result": {
          "type": "object",
          "properties": {
            "firstname": {
              "type": "string",
              "example": "Api"
            },
            "lastname": {
              "type": "string",
              "example": "Testing"
            },
            "email": {
              "type": "string",
              "example": "Traine@b&o.com"
            },
            "phone": {
              "type": "number",
              "example": 5551234567
            }
          }
        }
      }
    }
  }
}